Doc NetorkID & pagination params as int, not float

+++ python-packages/sra_client/src/zero_ex/sra_client/api/default_api.py
@@ -42,9 +42,9 @@ class DefaultApi(object):
         :param bool async_req: Whether request should be asynchronous.
         :param str asset_data_a: The assetData value for the first asset in the pair.
         :param str asset_data_b: The assetData value for the second asset in the pair.
-        :param float network_id: The id of the Ethereum network
-        :param float page: The number of the page to request in the collection.
-        :param float per_page: The number of records to return per page.
+        :param int network_id: The id of the Ethereum network
+        :param int page: The number of the page to request in the collection.
+        :param int per_page: The number of records to return per page.
 
         :return: :class:`RelayerApiAssetDataPairsResponseSchema`.
             If the method is called asynchronously returns the request thread.
@@ -71,9 +71,9 @@ class DefaultApi(object):
         :param bool async_req: Whether request should be asynchronous.
         :param str asset_data_a: The assetData value for the first asset in the pair.
         :param str asset_data_b: The assetData value for the second asset in the pair.
-        :param float network_id: The id of the Ethereum network
-        :param float page: The number of the page to request in the collection.
-        :param float per_page: The number of records to return per page.
+        :param int network_id: The id of the Ethereum network
+        :param int page: The number of the page to request in the collection.
+        :param int per_page: The number of records to return per page.
 
         :return: A tuple consisting of a
             :class:`RelayerApiAssetDataPairsResponseSchema`, an HTTP status
@@ -170,9 +170,9 @@ class DefaultApi(object):
         >>> result = thread.get() # doctest: +SKIP
 
         :param bool async_req: Whether request should be asynchronous.
-        :param float network_id: The id of the Ethereum network
-        :param float page: The number of the page to request in the collection.
-        :param float per_page: The number of records to return per page.
+        :param int network_id: The id of the Ethereum network
+        :param int page: The number of the page to request in the collection.
+        :param int per_page: The number of records to return per page.
 
         :return: :class:`RelayerApiFeeRecipientsResponseSchema`.
             If the method is called asynchronously, returns the request thread.
@@ -196,9 +196,9 @@ class DefaultApi(object):
         >>> result = thread.get() # doctest: +SKIP
 
         :param bool async_req: Whether request should be asynchronous.
-        :param float network_id: The id of the Ethereum network
-        :param float page: The number of the page to request in the collection.
-        :param float per_page: The number of records to return per page.
+        :param int network_id: The id of the Ethereum network
+        :param int page: The number of the page to request in the collection.
+        :param int per_page: The number of records to return per page.
 
         :return: A tuple consisting of a
             :class:`RelayerApiFeeRecipientsResponseSchema`, an HTTP status
@@ -281,7 +281,7 @@ class DefaultApi(object):
 
         :param bool async_req: Whether request should be asynchronous.
         :param str order_hash: The hash of the desired 0x order. (required)
-        :param float network_id: The id of the Ethereum network
+        :param int network_id: The id of the Ethereum network
 
         :return: :class:`RelayerApiOrderSchema`.
             If the method is called asynchronously, returns the request thread.
@@ -305,7 +305,7 @@ class DefaultApi(object):
 
         :param bool async_req: Whether request should be asynchronous.
         :param str order_hash: The hash of the desired 0x order. (required)
-        :param float network_id: The id of the Ethereum network
+        :param int network_id: The id of the Ethereum network
 
         :return: A tuple consisting of a
             :class:`RelayerApiOrderSchema`, an HTTP status code integer, and a
@@ -401,7 +401,7 @@ class DefaultApi(object):
         >>> result = thread.get() # doctest: +SKIP
 
         :param bool async_req: Whether request should be asynchronous.
-        :param float network_id: The id of the Ethereum network
+        :param int network_id: The id of the Ethereum network
         :param relayer_api_order_config_payload_schema: instance of
             :class:`RelayerApiOrderConfigPayloadSchema`. The fields of a 0x
             order the relayer may want to decide what configuration to send
@@ -436,7 +436,7 @@ class DefaultApi(object):
         >>> result = thread.get() # doctest: +SKIP
 
         :param bool async_req: Whether request should be asynchronous.
-        :param float network_id: The id of the Ethereum network
+        :param int network_id: The id of the Ethereum network
         :param relayer_api_order_config_payload_schema: instance of
             :class: `RelayerApiOrderConfigPayloadSchema`. The fields of a 0x
             order the relayer may want to decide what configuration to send
@@ -545,9 +545,9 @@ class DefaultApi(object):
         :param str quote_asset_data: assetData (makerAssetData or
             takerAssetData) designated as the quote currency in the currency
             pair calculation of price. (required)
-        :param float network_id: The id of the Ethereum network
-        :param float page: The number of the page to request in the collection.
-        :param float per_page: The number of records to return per page.
+        :param int network_id: The id of the Ethereum network
+        :param int page: The number of the page to request in the collection.
+        :param int per_page: The number of records to return per page.
 
         :return: :class:`RelayerApiOrderbookResponseSchema`.
             If the method is called asynchronously, returns the request thread.
@@ -594,9 +594,9 @@ class DefaultApi(object):
         :param str quote_asset_data: assetData (makerAssetData or
             takerAssetData) designated as the quote currency in the currency
             pair calculation of price. (required)
-        :param float network_id: The id of the Ethereum network
-        :param float page: The number of the page to request in the collection.
-        :param float per_page: The number of records to return per page.
+        :param int network_id: The id of the Ethereum network
+        :param int page: The number of the page to request in the collection.
+        :param int per_page: The number of records to return per page.
 
         :return: A tuple consisting of a
             :class:`RelayerApiOrderbookResponseSchema`, an HTTP status code
@@ -762,9 +762,9 @@ class DefaultApi(object):
             `0x Protocol v2 Specification
             <https://github.com/0xProject/0x-protocol-specification/blob/
             master/v2/v2-specification.md#order-message-format>`__
-        :param float network_id: The id of the Ethereum network
-        :param float page: The number of the page to request in the collection.
-        :param float per_page: The number of records to return per page.
+        :param int network_id: The id of the Ethereum network
+        :param int page: The number of the page to request in the collection.
+        :param int per_page: The number of records to return per page.
 
         :return: :class:`RelayerApiOrdersResponseSchema`.
             If the method is called asynchronously, returns the request thread.
@@ -839,9 +839,9 @@ class DefaultApi(object):
             `0x Protocol v2 Specification
             <https://github.com/0xProject/0x-protocol-specification/blob/
             master/v2/v2-specification.md#order-message-format>`__
-        :param float network_id: The id of the Ethereum network
-        :param float page: The number of the page to request in the collection.
-        :param float per_page: The number of records to return per page.
+        :param int network_id: The id of the Ethereum network
+        :param int page: The number of the page to request in the collection.
+        :param int per_page: The number of records to return per page.
 
         :return: A tuple consisting of a
             :class:`RelayerApiOrdersResponseSchema`, an HTTP status code
@@ -995,7 +995,7 @@ class DefaultApi(object):
         >>> result = thread.get() # doctest: +SKIP
 
         :param bool async_req: Whether request should be asynchronous.
-        :param float network_id: The id of the Ethereum network
+        :param int network_id: The id of the Ethereum network
         :param signed_order_schema: Instance of :class:`SignedOrderSchema`.
             A valid signed 0x order based on the schema.
 
@@ -1020,7 +1020,7 @@ class DefaultApi(object):
         >>> result = thread.get() # doctest: +SKIP
 
         :param bool async_req: Whether request should be asynchronous.
-        :param float network_id: The id of the Ethereum network
+        :param int network_id: The id of the Ethereum network
         :param signed_order_schema: Instance of :class:`SignedOrderSchema`
             A valid signed 0x order based on the schema.
